improve WC confirm sheet
